requirements.txt
Flask
flask-cors
pycryptodome
gunicorn
